Move netplay_get_architecture to frontend_driver.c
